requirements.txt
chromadb==0.4.24
sentence-transformers==5.1.1
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
python-dotenv
cryptography
groq
pandas
numpy<2.0
streamlit
